pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "OpenERF"
version = "0.6.1"
description = "One-line Effective Receptive Field extraction for pretrained timm models"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
    { name = "OpenERF Contributors" },
]
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "timm>=1.0.22",
    "numpy>=1.24",
    "Pillow>=10.0",
    "tqdm>=4.66",
    "matplotlib>=3.8",
]

[project.optional-dependencies]
fit = [
    "lmfit>=1.3",
    "scipy>=1.10",
]

[tool.setuptools.packages.find]
include = ["OpenERF*"]
